moved ShowSourceLine into an interface representing the development environment
