+++ BoostTestUi/DevEnv.h
@@ -1,0 +1,33 @@
+// (C) Copyright Gert-Jan de Vos 2012.
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt or copy at 
+// http://www.boost.org/LICENSE_1_0.txt)
+
+// See http://boosttestui.wordpress.com/ for the boosttestui home page.
+
+#ifndef BOOST_TESTUI_DEVENV_H
+#define BOOST_TESTUI_DEVENV_H
+
+#pragma once
+
+#include <string>
+
+namespace gj {
+
+class DevEnv
+{
+public:
+	DevEnv();
+	~DevEnv();
+
+	void ShowSourceLine(const std::string& fileName, int lineNr);
+
+private:
+	class Impl;
+
+	Impl* m_pImpl;
+};
+
+} // namespace gj
+
+#endif // BOOST_TESTUI_DEVENV_H
