--- BoostTestUi/ShowSourceLine.h
@@ -1,17 +1,0 @@
-// (C) Copyright Gert-Jan de Vos 2012.
-// Distributed under the Boost Software License, Version 1.0.
-// (See accompanying file LICENSE_1_0.txt or copy at 
-// http://www.boost.org/LICENSE_1_0.txt)
-
-// See http://boosttestui.wordpress.com/ for the boosttestui home page.
-
-#ifndef BOOST_TESTUI_SHOWSOURCELINE_H
-#define BOOST_TESTUI_SHOWSOURCELINE_H
-
-#pragma once
-
-#include <string>
-
-void ShowSourceLine(const std::string& fileName, int lineNr);
-
-#endif // BOOST_TESTUI_SHOWSOURCELINE_H
